Move ConstraintSystem and related traits to own crate, and move gm17 to own crate.
