+++ algebra/src/lib.rs
@@ -53,6 +53,10 @@ pub use self::fields::*;
 pub mod groups;
 pub use self::groups::*;
 
+
+mod to_field_vec;
+pub use to_field_vec::ToConstraintField;
+
 #[cfg(feature = "parallel")]
 pub mod msm;
 #[cfg(feature = "parallel")]
@@ -62,5 +66,3 @@ pub use self::msm::*;
 pub mod fft;
 #[cfg(feature = "parallel")]
 pub use self::fft::*;
-
-pub mod utils;
