+++ gm17/src/lib.rs
@@ -1,21 +1,39 @@
-use algebra::{PairingCurve, PairingEngine};
+//! An implementation of the [Groth-Maller][GM17] simulation extractable zkSNARK.
+//! [GM17]: https://eprint.iacr.org/2017/540
+#![deny(unused_import_braces, unused_qualifications, trivial_casts)]
+#![deny(trivial_numeric_casts, private_in_public, variant_size_differences)]
+#![deny(stable_features, unreachable_pub, non_shorthand_field_patterns)]
+#![deny(unused_attributes, unused_imports, unused_mut)]
+#![deny(renamed_and_removed_lints, stable_features, unused_allocation)]
+#![deny(unused_comparisons, bare_trait_objects, unused_must_use, const_err)]
 
-use crate::SynthesisError;
+#![forbid(unsafe_code)]
+
+#[macro_use]
+extern crate bench_utils;
+
+use algebra::{bytes::ToBytes, PairingCurve, PairingEngine};
+use r1cs_core::SynthesisError;
 use std::io::{self, Read, Result as IoResult, Write};
 
-use algebra::bytes::ToBytes;
+/// Reduce an R1CS instance to a *Square Arithmetic Program* instance.
+pub mod r1cs_to_sap;
 
-mod r1cs_to_sap;
+/// Generate public parameters for the GM17 zkSNARK construction.
+pub mod generator;
 
-mod generator;
-mod prover;
-mod verifier;
+/// Create proofs for the GM17 zkSNARK construction.
+pub mod prover;
+
+/// Verify proofs for the GM17 zkSNARK construction.
+pub mod verifier;
 
 #[cfg(test)]
 mod test;
 
 pub use self::{generator::*, prover::*, verifier::*};
 
+/// A proof in the GM17 SNARK.
 #[derive(Clone)]
 pub struct Proof<E: PairingEngine> {
     pub a: E::G1Affine,
@@ -49,17 +67,21 @@ impl<E: PairingEngine> Default for Proof<E> {
 }
 
 impl<E: PairingEngine> Proof<E> {
+    /// Serialize the proof into bytes, for storage on disk or transmission 
+    /// over the network.
     pub fn write<W: Write>(&self, mut _writer: W) -> io::Result<()> {
         // TODO: implement serialization
         unimplemented!()
     }
 
+    /// Deserialize the proof from bytes.
     pub fn read<R: Read>(mut _reader: R) -> io::Result<Self> {
         // TODO: implement serialization
         unimplemented!()
     }
 }
 
+/// A verification key in the GM17 SNARK.
 #[derive(Clone)]
 pub struct VerifyingKey<E: PairingEngine> {
     pub h_g2:       E::G2Affine,
@@ -109,17 +131,21 @@ impl<E: PairingEngine> PartialEq for VerifyingKey<E> {
 }
 
 impl<E: PairingEngine> VerifyingKey<E> {
+    /// Serialize the verification key into bytes, for storage on disk 
+    /// or transmission over the network.
     pub fn write<W: Write>(&self, mut _writer: W) -> io::Result<()> {
         // TODO: implement serialization
         unimplemented!()
     }
 
+    /// Deserialize the verification key from bytes.
     pub fn read<R: Read>(mut _reader: R) -> io::Result<Self> {
         // TODO: implement serialization
         unimplemented!()
     }
 }
 
+/// Full public (prover and verifier) parameters for the GM17 zkSNARK.
 #[derive(Clone)]
 pub struct Parameters<E: PairingEngine> {
     pub vk:           VerifyingKey<E>,
@@ -150,17 +176,21 @@ impl<E: PairingEngine> PartialEq for Parameters<E> {
 }
 
 impl<E: PairingEngine> Parameters<E> {
+    /// Serialize the parameters to bytes.
     pub fn write<W: Write>(&self, mut _writer: W) -> io::Result<()> {
         // TODO: implement serialization
         unimplemented!()
     }
 
+    /// Deserialize the public parameters from bytes.
     pub fn read<R: Read>(mut _reader: R, _checked: bool) -> io::Result<Self> {
         // TODO: implement serialization
         unimplemented!()
     }
 }
 
+/// Preprocessed verification key parameters that enable faster verification
+/// at the expense of larger size in memory.
 #[derive(Clone)]
 pub struct PreparedVerifyingKey<E: PairingEngine> {
     pub vk:                VerifyingKey<E>,
