+++ gm17/src/verifier.rs
