+++ r1cs-core/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "r1cs-core"
+version = "0.1.0"
+authors = [
+    "Sean Bowe",
+    "Alessandro Chiesa",
+    "Matthew Green",
+    "Ian Miers",
+    "Pratyush Mishra",
+    "Howard Wu"
+]
+description = "A library for rank-one constraint systems"
+homepage = "https://libzexe.org"
+repository = "https://github.com/scipr/zexe"
+documentation = "https://docs.rs/r1cs-core/"
+keywords = ["zero knowledge", "cryptography", "zkSNARK", "SNARK", "constraint systems"]
+categories = ["cryptography"]
+include = ["Cargo.toml", "src", "README.md", "LICENSE-APACHE", "LICENSE-MIT"]
+license = "MIT/Apache-2.0"
+edition = "2018"
+
+[dependencies]
+algebra = { path = "../algebra" }
+smallvec = { version = "0.6" }
