+++ r1cs-core/src/constraint_system.rs
@@ -1,0 +1,232 @@
+use std::marker::PhantomData;
+use algebra::Field;
+
+use crate::{Index, Variable, LinearCombination, SynthesisError};
+
+/// Represents a constraint system which can have new variables
+/// allocated and constrains between them formed.
+pub trait ConstraintSystem<F: Field>: Sized {
+    /// Represents the type of the "root" of this constraint system
+    /// so that nested namespaces can minimize indirection.
+    type Root: ConstraintSystem<F>;
+
+    /// Return the "one" input variable
+    fn one() -> Variable {
+        Variable::new_unchecked(Index::Input(0))
+    }
+
+    /// Allocate a private variable in the constraint system. The provided
+    /// function is used to determine the assignment of the variable. The
+    /// given `annotation` function is invoked in testing contexts in order
+    /// to derive a unique name for this variable in the current namespace.
+    fn alloc<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>;
+
+    /// Allocate a public variable in the constraint system. The provided
+    /// function is used to determine the assignment of the variable.
+    fn alloc_input<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>;
+
+    /// Enforce that `A` * `B` = `C`. The `annotation` function is invoked in
+    /// testing contexts in order to derive a unique name for the constraint
+    /// in the current namespace.
+    fn enforce<A, AR, LA, LB, LC>(&mut self, annotation: A, a: LA, b: LB, c: LC)
+    where
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+        LA: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LB: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LC: FnOnce(LinearCombination<F>) -> LinearCombination<F>;
+
+    /// Create a new (sub)namespace and enter into it. Not intended
+    /// for downstream use; use `namespace` instead.
+    fn push_namespace<NR, N>(&mut self, name_fn: N)
+    where
+        NR: Into<String>,
+        N: FnOnce() -> NR;
+
+    /// Exit out of the existing namespace. Not intended for
+    /// downstream use; use `namespace` instead.
+    fn pop_namespace(&mut self);
+
+    /// Gets the "root" constraint system, bypassing the namespacing.
+    /// Not intended for downstream use; use `namespace` instead.
+    fn get_root(&mut self) -> &mut Self::Root;
+
+    /// Begin a namespace for this constraint system.
+    fn ns<'a, NR, N>(&'a mut self, name_fn: N) -> Namespace<'a, F, Self::Root>
+    where
+        NR: Into<String>,
+        N: FnOnce() -> NR,
+    {
+        self.get_root().push_namespace(name_fn);
+
+        Namespace(self.get_root(), PhantomData)
+    }
+
+    /// Output the number of constraints in the system.
+    fn num_constraints(&self) -> usize;
+}
+
+/// This is a "namespaced" constraint system which borrows a constraint system
+/// (pushing a namespace context) and, when dropped, pops out of the namespace
+/// context.
+pub struct Namespace<'a, F: Field, CS: ConstraintSystem<F>>(&'a mut CS, PhantomData<F>);
+
+/// Computations are expressed in terms of rank-1 constraint systems (R1CS).
+/// The `generate_constraints` method is called to generate constraints for
+/// both CRS generation and for proving.
+pub trait ConstraintSynthesizer<F: Field> {
+    /// Drives generation of new constraints inside `CS`.
+    fn generate_constraints<CS: ConstraintSystem<F>>(self, cs: &mut CS) -> Result<(), SynthesisError>;
+}
+
+
+impl<F: Field, CS: ConstraintSystem<F>> ConstraintSystem<F> for Namespace<'_, F, CS> {
+    type Root = CS::Root;
+
+    #[inline]
+    fn one() -> Variable {
+        CS::one()
+    }
+
+    #[inline]
+    fn alloc<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+    {
+        self.0.alloc(annotation, f)
+    }
+
+    #[inline]
+    fn alloc_input<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+    {
+        self.0.alloc_input(annotation, f)
+    }
+
+    #[inline]
+    fn enforce<A, AR, LA, LB, LC>(&mut self, annotation: A, a: LA, b: LB, c: LC)
+    where
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+        LA: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LB: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LC: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+    {
+        self.0.enforce(annotation, a, b, c)
+    }
+
+    // Downstream users who use `namespace` will never interact with these
+    // functions and they will never be invoked because the namespace is
+    // never a root constraint system.
+
+    #[inline]
+    fn push_namespace<NR, N>(&mut self, _: N)
+    where
+        NR: Into<String>,
+        N: FnOnce() -> NR,
+    {
+        panic!("only the root's push_namespace should be called");
+    }
+
+    #[inline]
+    fn pop_namespace(&mut self) {
+        panic!("only the root's pop_namespace should be called");
+    }
+
+    #[inline]
+    fn get_root(&mut self) -> &mut Self::Root {
+        self.0.get_root()
+    }
+
+    #[inline]
+    fn num_constraints(&self) -> usize {
+        self.0.num_constraints()
+    }
+}
+
+impl<F: Field, CS: ConstraintSystem<F>> Drop for Namespace<'_, F, CS> {
+    #[inline]
+    fn drop(&mut self) {
+        self.get_root().pop_namespace()
+    }
+}
+
+/// Convenience implementation of ConstraintSystem<F> for mutable references to
+/// constraint systems.
+impl<F: Field, CS: ConstraintSystem<F>> ConstraintSystem<F> for &mut CS {
+    type Root = CS::Root;
+
+    #[inline]
+    fn one() -> Variable {
+        CS::one()
+    }
+
+    #[inline]
+    fn alloc<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+    {
+        (**self).alloc(annotation, f)
+    }
+
+    #[inline]
+    fn alloc_input<FN, A, AR>(&mut self, annotation: A, f: FN) -> Result<Variable, SynthesisError>
+    where
+        FN: FnOnce() -> Result<F, SynthesisError>,
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+    {
+        (**self).alloc_input(annotation, f)
+    }
+
+    #[inline]
+    fn enforce<A, AR, LA, LB, LC>(&mut self, annotation: A, a: LA, b: LB, c: LC)
+    where
+        A: FnOnce() -> AR,
+        AR: Into<String>,
+        LA: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LB: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+        LC: FnOnce(LinearCombination<F>) -> LinearCombination<F>,
+    {
+        (**self).enforce(annotation, a, b, c)
+    }
+
+    #[inline]
+    fn push_namespace<NR, N>(&mut self, name_fn: N)
+    where
+        NR: Into<String>,
+        N: FnOnce() -> NR,
+    {
+        (**self).push_namespace(name_fn)
+    }
+
+    #[inline]
+    fn pop_namespace(&mut self) {
+        (**self).pop_namespace()
+    }
+
+    #[inline]
+    fn get_root(&mut self) -> &mut Self::Root {
+        (**self).get_root()
+    }
+
+    #[inline]
+    fn num_constraints(&self) -> usize {
+        (**self).num_constraints()
+    }
+}
