+++ r1cs-lib/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "r1cs-stdlib"
+version = "0.1.0"
+authors = ["Pratyush Mishra <pratyushmishra@berkeley.edu>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
